test: move API tests from basic.rs to mod.rs
